added checks for path

+++ Tools/MDRemoveOldFile/Remove-MDOldFile/Remove-MDOldFile.psm1
@@ -2,21 +2,27 @@ function Remove-MDOldFile {
     [CmdletBinding()]
     param (
         # Path of directory
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [string]
         $Path,
         # Number of days to keep
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [int]
-        $NumberOfDaysToKeep = 7
+        $NumberOfDaysToKeep
     )
     
     begin {
     }
     
     process {
-        Get-ChildItem -Path $Path | Where-Object -Property 'LastWriteTime' -LE (Get-Date).AddDays( -$NumberOfDaysToKeep) |
-        Remove-Item -Force
+        if ( Test-Path $Path) {
+            Get-ChildItem -Path $Path | Where-Object -Property 'LastWriteTime' -LT (Get-Date).AddDays(-$NumberOfDaysToKeep) |
+            Remove-Item -Force    
+        } else {
+            Write-Warning "Path not found"
+        }
     }
     
     end {
